Add Currency Selection UI Components
feat: create SelectCurrency SwiftUI view

Created the CurrencyIcon SwiftUI view, which displays a currency image with accompanying text. This view is designed for use in the LOTR Converter app.

feat: create CurrencyIcon SwiftUI view

Created the CurrencyIcon SwiftUI view, which displays a currency image with accompanying text. This view is designed for use in the LOTR Converter app.